ext4 crypto: implement the ext4 encryption write path
When encryptin, we send one encrypted bounce page at a time out to the
block layer. This isn't the most efficient approach, but it's
relatively low-impact and can be improved later if need be.  There's
talk amongst the ext4 maintainers around some day getting rid of
buffer heads, and the encryption path can fold into that work at the
time it's done.

Pulls block_write_begin() into fs/ext4/inode.c because it might need
to do a low-level read of the existing data, in which case we need to
decrypt it.

Signed-off-by: Michael Halcrow <mhalcrow@google.com>
Signed-off-by: Ildar Muslukhov <ildarm@google.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>